Lib: Picker: JLPCB: Move pickers, add can_query trait
Add can_query_jlcpcb_db trait
Move pickers from libs/picker/jlcpcb.py to libs/picker/jlcpcb/pickers.py
Move remaining jlcpcb.py to libs.picker/jlcpcb/jlcpcb.py
Update test
